Rename fra .server til .util
