+++ app/models/behandling.server.ts
@@ -2,7 +2,7 @@ import createClient from "openapi-fetch";
 
 import { getBehandlingOboToken } from "~/utils/auth.utils.server";
 import { getEnv } from "~/utils/env.utils";
-import { handleHttpProblem } from "~/utils/error-response.server";
+import { handleHttpProblem } from "~/utils/error-response.utils";
 import { getHeaders } from "~/utils/fetch.utils";
 
 import { paths } from "../../openapi/behandling-typer";
