+++ app/server-side-actions/kvitter-avklaring-action.ts
@@ -2,7 +2,7 @@ import invariant from "tiny-invariant";
 
 import { IAlert } from "~/context/alert-context";
 import { kvitterAvklaring } from "~/models/behandling.server";
-import { getHttpProblemAlert } from "~/utils/error-response.server";
+import { getHttpProblemAlert } from "~/utils/error-response.utils";
 
 export async function kvitterAvklaringAction(request: Request, formData: FormData) {
   const behandlingId = formData.get("behandling-id") as string;
