+++ app/server-side-actions/lagre-opplysning-action.ts
@@ -2,7 +2,7 @@ import invariant from "tiny-invariant";
 
 import { IAlert } from "~/context/alert-context";
 import { endreOpplysning } from "~/models/behandling.server";
-import { getHttpProblemAlert } from "~/utils/error-response.server";
+import { getHttpProblemAlert } from "~/utils/error-response.utils";
 
 export async function lagreOpplysningAction(request: Request, formData: FormData) {
   const behandlingId = formData.get("behandlingId") as string;
