+++ app/server-side-actions/legg-tilbake-oppgave-action.ts
@@ -3,7 +3,7 @@ import { ActionFunctionArgs, redirect } from "react-router";
 import { IAlert } from "~/context/alert-context";
 import { leggTilbakeOppgave } from "~/models/saksbehandling.server";
 import { commitSession, getSession } from "~/sessions";
-import { getHttpProblemAlert } from "~/utils/error-response.server";
+import { getHttpProblemAlert } from "~/utils/error-response.utils";
 
 export async function leggTilbakeOppgaveAction(
   request: Request,
