+++ app/server-side-actions/send-til-arena-action.ts
@@ -3,7 +3,7 @@ import { ActionFunctionArgs, redirect } from "react-router";
 import { IAlert } from "~/context/alert-context";
 import { avbrytBehandling } from "~/models/behandling.server";
 import { commitSession, getSession } from "~/sessions";
-import { getHttpProblemAlert } from "~/utils/error-response.server";
+import { getHttpProblemAlert } from "~/utils/error-response.utils";
 
 export async function sendTilArenaAction(
   request: Request,
