+++ app/server-side-actions/send-til-kontroll-action.ts
@@ -4,7 +4,7 @@ import invariant from "tiny-invariant";
 import { IAlert } from "~/context/alert-context";
 import { sendOppgaveTilKontroll } from "~/models/saksbehandling.server";
 import { commitSession, getSession } from "~/sessions";
-import { getHttpProblemAlert } from "~/utils/error-response.server";
+import { getHttpProblemAlert } from "~/utils/error-response.utils";
 
 export async function sendTilKontrollAction(
   request: Request,
